added testPlanVersionsResolver resolver
